requirements.txt
streamlit==1.28.1
groq==0.4.2
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.0.350
langchain-groq==0.0.1
langchain-community==0.0.10 